Add a more helpful error message to `oneOf` (#25)
This changes the `oneOf` error message to be much more descriptive of the actual issue when all given decoders fail, and includes the reasons each of them failed for better debugging. This comes at a performance cost, but it seems like a negligible cost for a significant benefit in the most common cases.